refactor(tests): move policy reports tests alongside source files
