+++ frontend/src/utilities/policyReports/getTeams.test.js
@@ -1,42 +1,37 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { fetchDatasetRepositoriesForUser } from '../../../src/utilities/policyReports/getRepositories';
+import { fetchDatasetTeamsForUser } from './getTeams';
 
-vi.mock('../../../src/utilities/customFetch', () => ({
+vi.mock('../customFetch', () => ({
   default: vi.fn(),
 }));
 
-import customFetch from '../../../src/utilities/customFetch';
+import customFetch from '../customFetch';
 
-describe('fetchDatasetRepositoriesForUser', () => {
+describe('fetchDatasetTeamsForUser', () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
 
-  it('returns repositories for a given organisation and dataset', async () => {
-    const mockRepos = ['repo-a', 'repo-b', 'repo-c'];
-    customFetch.mockResolvedValue({
-      json: async () => ({ repositories: mockRepos }),
-    });
+  it('returns teams for a given organisation and dataset', async () => {
+    const mockTeams = ['team-alpha', 'team-beta'];
+    customFetch.mockResolvedValue({ json: async () => ({ teams: mockTeams }) });
 
-    const result = await fetchDatasetRepositoriesForUser(
+    const result = await fetchDatasetTeamsForUser(
       'ONS-Innovation',
       '2024-01-15T10:00:00Z'
     );
 
     expect(customFetch).toHaveBeenCalledWith(
-      expect.stringContaining('/policy-reports/api/repositories'),
+      expect.stringContaining('/policy-reports/api/teams'),
       { credentials: 'include' }
     );
-    expect(result).toEqual(mockRepos);
+    expect(result).toEqual(mockTeams);
   });
 
   it('URL-encodes the organisation and dataset parameters', async () => {
-    customFetch.mockResolvedValue({ json: async () => ({ repositories: [] }) });
+    customFetch.mockResolvedValue({ json: async () => ({ teams: [] }) });
 
-    await fetchDatasetRepositoriesForUser(
-      'ONS-Innovation',
-      '2024-01-15T10:00:00Z'
-    );
+    await fetchDatasetTeamsForUser('ONS-Innovation', '2024-01-15T10:00:00Z');
 
     const calledUrl = customFetch.mock.calls[0][0];
     expect(calledUrl).toContain('organisation=ONS-Innovation');
@@ -45,10 +40,10 @@ describe('fetchDatasetRepositoriesForUser', () => {
     );
   });
 
-  it('returns an empty array when repositories key is missing from response', async () => {
+  it('returns an empty array when teams key is missing from response', async () => {
     customFetch.mockResolvedValue({ json: async () => ({}) });
 
-    const result = await fetchDatasetRepositoriesForUser(
+    const result = await fetchDatasetTeamsForUser(
       'ONS-Innovation',
       '2024-01-15T10:00:00Z'
     );
@@ -57,17 +52,14 @@ describe('fetchDatasetRepositoriesForUser', () => {
   });
 
   it('returns an empty array when organisation is falsy', async () => {
-    const result = await fetchDatasetRepositoriesForUser(
-      '',
-      '2024-01-15T10:00:00Z'
-    );
+    const result = await fetchDatasetTeamsForUser('', '2024-01-15T10:00:00Z');
 
     expect(customFetch).not.toHaveBeenCalled();
     expect(result).toEqual([]);
   });
 
   it('returns an empty array when dataset is falsy', async () => {
-    const result = await fetchDatasetRepositoriesForUser('ONS-Innovation', '');
+    const result = await fetchDatasetTeamsForUser('ONS-Innovation', '');
 
     expect(customFetch).not.toHaveBeenCalled();
     expect(result).toEqual([]);
@@ -76,7 +68,7 @@ describe('fetchDatasetRepositoriesForUser', () => {
   it('returns an empty array when the fetch throws', async () => {
     customFetch.mockRejectedValue(new Error('Network error'));
 
-    const result = await fetchDatasetRepositoriesForUser(
+    const result = await fetchDatasetTeamsForUser(
       'ONS-Innovation',
       '2024-01-15T10:00:00Z'
     );
